rename 'changelog' modules to 'cypher'
- rename cypher-generating modules and scripts from 'changelog' to 'cypher'
- separate 'changelogs.py' into modules for models, cdes, and utils
- formatting
