+++ pyproject.toml
@@ -41,3 +41,8 @@ dev = [
     "pytest>=8.3.4",
     "testcontainers>=4.9.2",
 ]
+
+[tool.pytest.ini_options]
+markers = [
+    "docker: marks tests that require docker",
+]
