+++ scripts/make_model_changelog.py
@@ -13,7 +13,7 @@ from pathlib import Path
 import click
 from bento_mdf.mdf import MDF
 
-from bento_mdb_updates.changelogs import ModelToChangelogConverter
+from bento_mdb_updates.model_cypher import ModelToChangelogConverter
 
 logger = logging.getLogger(__name__)
