+++ tests/test_cde_cypher.py
@@ -1,0 +1,4 @@
+class TestCDEChangelog:
+    """Tests for cde changelog generation script."""
+
+
