+++ tests/test_cypher_utils.py
@@ -1,8 +1,8 @@
-"""Tests for changelog utilities."""
+"""Tests for cypher utilities."""
 
 from bento_meta.objects import Property
 
-from bento_mdb_updates.changelogs import escape_quotes_in_attr
+from bento_mdb_updates.cypher_utils import escape_quotes_in_attr
 
 
 def test_escape_quotes_in_attr() -> None:
